requirements.txt
pydantic==1.4
uvicorn==0.11.3
gunicorn==20.0.4
fastapi==0.52.0
python-dotenv==0.15.0
fastapi-cache2==0.1.8
requests~=2.28.2